Consolidate some of the duplicate code
