+++ src/gui/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(GUI_SOURCE
     audio_gui.cpp
     mesh_gui.cpp
+    mesh_manager.cpp
     circular_mesh_manager.cpp
     rectangular_mesh_manager.cpp
     )
