+++ src/mesh_graph/trimesh_test.cpp
@@ -64,6 +64,7 @@ int main()
     info.use_nonlinear_allpass = false;
     info.nonlinear_allpass_coeffs[0] = 0.5f;
     info.nonlinear_allpass_coeffs[1] = 0.1f;
+    info.get_rimguide_pos = std::bind(get_boundary_position, kRadius, std::placeholders::_1);
 
     TriMesh mesh(kGridX, kGridY, sample_distance);
     auto mask = mesh.get_mask_for_radius(max_radius);
